refactor: extract props to dedicated files
