+++ tests/unit/components/LGeoJson.test.ts
@@ -8,30 +8,10 @@ import {
 } from '@/tests/helper/propsBindingTests'
 import { testEmitsReady } from '@/tests/helper/emitTests'
 import { mockAddLayer, mockRemoveLayer, testAddLayer } from '@/tests/helper/injectionsTests'
-import { GeoJSON, Layer } from 'leaflet'
+import { GeoJSON } from 'leaflet'
 import LGeoJson from '@/components/LGeoJson.vue'
 import geoJson from './geo.json'
-import geoJsonReplace from './geo_replace.json'
-import { mergeReactiveProps } from '@/tests/helper/props'
-import { layerGroupProps } from './wrapper/LLayerGroup'
-
-const geoJsonProps = mergeReactiveProps(layerGroupProps, {
-    geojson: geoJsonReplace,
-    optionsStyle: () => ({
-        opacity: 0.35
-    }),
-    expecting: {
-        ...layerGroupProps.expecting,
-        geojson: (leafletObject: GeoJSON) => {
-            expect(leafletObject.toGeoJSON()).toStrictEqual(geoJsonReplace)
-        },
-        optionsStyle: (leafletObject: GeoJSON) => {
-            leafletObject.eachLayer((layer: Layer) => {
-                expect(layer.options).toStrictEqual({ opacity: 0.35 })
-            })
-        }
-    }
-})
+import { geoJsonProps } from './wrapper/LGeoJson'
 
 const createWrapper = async (props = {}) => {
     const wrapper = shallowMount(LGeoJson, {
