+++ tests/unit/components/wrapper/LCircle.ts
@@ -1,0 +1,6 @@
+import { mergeReactiveProps } from '@/tests/helper/props'
+import { circleMarkerProps } from './LCircleMarker'
+
+export const circleProps = mergeReactiveProps(circleMarkerProps, {
+    radius: 10000
+})
