+++ tests/unit/components/wrapper/LControlAttribution.ts
@@ -1,0 +1,6 @@
+import { mergeReactiveProps } from '@/tests/helper/props'
+import { controlAbstractProps } from './LControl'
+
+export const controlAttributionProps = mergeReactiveProps(controlAbstractProps, {
+    prefix: 'new prefix'
+})
