+++ tests/unit/components/wrapper/LControlScale.ts
@@ -1,0 +1,4 @@
+import { mergeReactiveProps } from '@/tests/helper/props'
+import { controlAbstractProps } from './LControl'
+
+export const controlScaleProps = mergeReactiveProps(controlAbstractProps, {})
