+++ tests/unit/components/wrapper/LFeatureGroup.ts
@@ -1,0 +1,4 @@
+import { mergeReactiveProps } from '@/tests/helper/props'
+import { layerGroupProps } from './LLayerGroup'
+
+export const featureGroupProps = mergeReactiveProps(layerGroupProps, {})
