+++ tests/unit/components/wrapper/LSVGOverlay.ts
@@ -1,0 +1,4 @@
+import { mergeReactiveProps } from '@/tests/helper/props'
+import { imageOverlayAbstractProps } from './LImageOverlay'
+
+export const svgOverlayProps = mergeReactiveProps(imageOverlayAbstractProps, {})
