+++ tests/unit/components/wrapper/LTooltip.ts
@@ -1,0 +1,4 @@
+import { mergeReactiveProps } from '@/tests/helper/props'
+import { popperProps } from '@/tests/helper/propsBindingTests'
+
+export const tooltipProps = mergeReactiveProps(popperProps, {})
