+++ tests/unit/components/wrapper/LWmsTileLayer.ts
@@ -1,0 +1,13 @@
+import { mergeReactiveProps } from '@/tests/helper/props'
+import { tileLayerProps } from './LTileLayer'
+import { TileLayer } from 'leaflet'
+import { expect } from 'vitest'
+
+export const wmsTileLayerProps = mergeReactiveProps(tileLayerProps, {
+    url: 'https://ows.terrestris.de/osm/service?',
+    expecting: {
+        url: (l: TileLayer & { _url: string }) => {
+            expect(l._url).toBe(wmsTileLayerProps.url)
+        }
+    }
+})
